Role Guard and design (Angular Material)
Adding role guard.
New design for components.

+++ ClientApp/src/app/animal/list-animal/list-animal.component.ts
@@ -11,7 +11,7 @@ import { AddAnimalComponent } from '../add-animal/add-animal.component';
 })
 export class ListAnimalComponent implements OnInit {
 
-  animals: AnimalDto[] = [];  
+  animals: AnimalDto[] = [];
   page = 1;
   pageSize = 8;
 
@@ -19,21 +19,30 @@ export class ListAnimalComponent implements OnInit {
 
   open() {
     const modalRef = this.modalService.open(AddAnimalComponent);
-    modalRef.result.then(() => this.refreshAnimals())
+    modalRef.result.then(() => this.refreshAnimals(), () => { })
   }
 
   ngOnInit(): void {
     this.refreshAnimals();
   }
 
-  deleteAnimal(id: number) { 
-    this.animalService.deleteAnimal(id).subscribe(() => this.refreshAnimals());
+  deleteAnimal(id: number) {
+    if (confirm('Biztosan szeretnéd törölni?')) {
+      this.animalService.deleteAnimal(id).subscribe(() => this.refreshAnimals());
+    }
   }
 
-  refreshAnimals(){
+  refreshAnimals() {
     this.animalService.getUserAnimals().subscribe((animals: AnimalDto[]) => {
       this.animals = animals;
     })
   }
 
 }
+
+/*
+<mat-icon style="clip-path: inset(0% 50% 0% 0%);">wc</mat-icon> <!-- Male -->
+<mat-icon style="clip-path: inset(0% 0% 0% 50%);">wc</mat-icon> <!-- Female -->
+
+
+*/
